Properly sort columns with dates in listviews
Checks if the column header Tag is a DateTime object and if so, sorts the column by date. This approach prevents performance impact when sorting columns that are not dates and allows any user of MyListView to tag a column to be sorted by date. Supports expansion to other sorting formats.
Tested with history of 80,000 items.

Closes #1685