Don't generate direct download links when PreviewLink is checked
See issue: https://github.com/ShareX/ShareX/issues/3868
Not adding either /preview or /download to URL generates links that show a preview of a file. /preview only works for images (showing literally only the image in the browser). This change makes sure the logic is right.

Note: The logic is right so it should work. But I'm not sure if the syntax is right.