Update color picker format options to include 0-1 colors
I used capital letters R, G, and B since they had not been used yet. This format of colors is used with a number of things, most notably Unity, and this will save me and hopefully others from having to convert normal colors to 0-1 range every so often, which I say is a win.
I have it set to round to 3 decimal places, but that's totally arbitrary and any precision will work, although some level of rounding seems sensible. I could probably make it so users can specify their own rounding precision, however this would be somewhat more effort to implement and would be a more significant deviation from the rest of the formatting rules, so I'll only add that if someone else says it'll actually be accepted. 
Also, why is banker rounding the default smh c # is is dumb